chore: schema 분리
